+++ doapi/doapi.py
@@ -38,7 +38,7 @@ class doapi(object):
     DEFAULT_ENDPOINT = 'https://api.digitalocean.com'
 
     def __init__(self, api_token, endpoint=DEFAULT_ENDPOINT, timeout=None,
-                 wait_interval=5, wait_time=None, per_page=None):
+                 wait_interval=2, wait_time=None, per_page=None):
         #: The API token used for authentication
         self.api_token = api_token
         #: The API endpoint URL relative to which requests will be made