+++ doapi/__init__.py
@@ -10,7 +10,7 @@ from .image       import Image
 from .ssh_key     import SSHKey
 from .tag         import Tag
 
-__version__      = '0.2.0.dev1'
+__version__      = '0.2.0'
 __author__       = 'John Thorvald Wodder II'
 __author_email__ = 'doapi@varonathe.org'
 __license__      = 'MIT'