+++ CHANGELOG.rst
@@ -7,6 +7,8 @@ v0.2.0 (in development)
 - **Breaking**: ``doapi.wait_droplets`` and ``Droplet.wait`` no longer have the
   option to wait for droplets' most recent actions to complete
 - **Breaking**: ``--multiple`` now also matches by ID, slug, & fingerprint
+- **Bugfix**: The ``--wait-interval`` and ``--wait-time`` options now actually
+  have an effect
 - New methods:
 
     - ``doapi.wait_actions_on_objects``
@@ -20,10 +22,14 @@ v0.2.0 (in development)
 - Removed droplet upgrade code, as the API no longer supports it
 - Gave ``Droplet`` ``ipv4_address`` and ``ipv6_address`` attributes
 - Decreased the default wait interval to 2 seconds
+- Handle fetching the last action on a resource that has never been acted on
+- The ``wait_for_action`` methods now return ``None`` if the resource no longer
+  exists by the time the action completes
 
 v0.1.1 (2016-04-29)
 -------------------
-- Fixed a bug with creating droplets with SSH keys from the command line
+- **Bugfix**: Fixed a bug with creating droplets with SSH keys from the command
+  line
 
 v0.1.0 (2016-04-07)
 -------------------